+++ src/controllers/ticket.controller.ts
@@ -57,6 +57,9 @@ export default {
   async findOne(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+      if (!isValidObjectId(id)) {
+        return response.notFound(res, "Not found!");
+      }
       const result = await TicketModel.findById(id);
       if (!result) {
         return response.notFound(res, "Not found!");
@@ -69,6 +72,9 @@ export default {
   async update(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+      if (!isValidObjectId(id)) {
+        return response.notFound(res, "Not found!");
+      }
       const result = await TicketModel.findByIdAndUpdate(id, req.body, {
         new: true,
       });
@@ -80,6 +86,9 @@ export default {
   async remove(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+      if (!isValidObjectId(id)) {
+        return response.notFound(res, "Not found!");
+      }
       const result = await TicketModel.findByIdAndDelete(id, {
         new: true,
       });
@@ -92,7 +101,7 @@ export default {
     try {
       const { eventId } = req.params;
       if (!isValidObjectId(eventId)) {
-        return response.error(res, null, "Tickets Not Found!");
+        return response.notFound(res, "Tickets Not Found!");
       }
       const result = await TicketModel.find({ events: eventId }).exec();
       response.success(res, result, "Tickets found for this event!");
